# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This server controls computer craft turtles.",
        "title": "Turtle C2 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/command": {
            "post": {
                "description": "send command to turtle",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "run command",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "command to send",
                        "name": "command",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Command"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response from the turtle",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/disconnect": {
            "delete": {
                "description": "disconnect turtle",
                "tags": [
                    "session"
                ],
                "summary": "disconnect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response from the turtle",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions": {
            "get": {
                "description": "get active sessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "get active sessions",
                "responses": {
                    "200": {
                        "description": "array of connected turtle IDs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Command": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                }
            }
        }
    }
}